requirements.txt
torch
streamlit
transformers
langchain
faiss-cpu
sentense-transformers